Fix reference to browserconfig.xml
This small patch fixes an issue with `browserconfig.xml`, linked-to by `<meta name="msapplication-config" ...>`. Currently, because it uses the `content` attribute to reference its path, it isn't rewritten to respect `prefix` and `publicPath` as-configured in the `favicons-webpack-plugin` module.

This changeset applies the same `href=` rewrites on the icons, to the `browserconfig.xml` - but only for tags containing `msapplication`, to avoid breaking any other behavior.